ai_service/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
web3
eth-account
eth-hash[pycryptodome]
